mcp_components/pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atlassian-python-api>=4.0.7",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
